+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### IMPROVEMENTS / OPTIMIZATIONS
 
+* All EmbeddedMortarProject.deployToMortar() methods now return the git hash of the related commit.
+
 ### BUG FIXES
 
 ## Release 0.1.0